add tests for the Form and FormContainer
